requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic[email]
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
requests==2.31.0
jinja2==3.1.3
unstructured==0.11.0
unstructured[docx,pdf,pptx,images]==0.11.0
llama-cpp-python==0.2.19
sentence-transformers==2.2.2
pillow==10.2.0
duckduckgo-search==3.9.9
beautifulsoup4==4.12.2
boto3==1.34.41